♻️ Move the base test suite into base test module.
