+++ test/unit_test/model/rest_api_doc_config/_base_config.py
@@ -1,9 +1,16 @@
 import logging
 import re
 from abc import ABC, ABCMeta, abstractmethod
-from typing import List, Union
+from collections import namedtuple
+from typing import List, Optional, Type, Union
 
 import pytest
+
+from pymock_server.model.api_config.apis._property import BaseProperty
+from pymock_server.model.api_config.format import Format
+from pymock_server.model.api_config.value import FormatStrategy, ValueFormat
+from pymock_server.model.api_config.variable import Size, Variable
+from pymock_server.model.rest_api_doc_config._js_handlers import ApiDocValueFormat
 
 # isort: off
 
@@ -18,6 +25,7 @@ from pymock_server.model.rest_api_doc_config._base import (
     set_openapi_version,
 )
 from pymock_server.model.rest_api_doc_config._base_model_adapter import (
+    BasePropertyDetailAdapter,
     BaseRequestParameterAdapter,
 )
 from pymock_server.model.rest_api_doc_config._model_adapter import (
@@ -746,3 +754,212 @@ class _OpenAPIDocumentDataModelTestSuite(metaclass=ABCMeta):
     @abstractmethod
     def _verify_api_config_model(self, under_test: _Config, data_from: Transferable) -> None:
         pass
+
+
+def _adapter_config(
+    required: bool,
+    value_type: str,
+    name: Optional[str] = None,
+    format: Optional[FormatAdapter] = None,
+    items: Optional[List[dict]] = None,
+) -> dict:
+    return {
+        "name": name,
+        "required": required,
+        "value_type": value_type,
+        "format": format,
+        "items": items,
+    }
+
+
+ExpectModelProps = namedtuple("ExpectModelProps", ("name", "required", "value_type", "format", "items"))
+
+
+class _BasePropertyDetailAdapterTestSuite(metaclass=ABCMeta):
+    @abstractmethod
+    @pytest.fixture(scope="function")
+    def adapter_model_obj(self) -> Type[BasePropertyDetailAdapter]:
+        pass
+
+    @abstractmethod
+    @pytest.fixture(scope="function")
+    def pymock_model(self) -> BaseProperty:
+        pass
+
+    @pytest.mark.parametrize(
+        ("under_test_data", "expect"),
+        [
+            (
+                _adapter_config(name="param", required=True, value_type="str"),
+                ExpectModelProps(name="param", required=True, value_type="str", format=None, items=None),
+            ),
+            (
+                _adapter_config(name="param", required=False, value_type="int"),
+                ExpectModelProps(name="param", required=False, value_type="int", format=None, items=None),
+            ),
+            (
+                _adapter_config(name="param", required=False, value_type="list"),
+                ExpectModelProps(name="param", required=False, value_type="list", format=None, items=None),
+            ),
+            (
+                _adapter_config(
+                    name="param",
+                    required=True,
+                    value_type="str",
+                    format=FormatAdapter(formatter=ApiDocValueFormat.Int32),
+                ),
+                ExpectModelProps(
+                    name="param",
+                    required=True,
+                    value_type="str",
+                    format=Format(
+                        strategy=FormatStrategy.BY_DATA_TYPE,
+                        size=Size(max_value=9223372036854775807, min_value=-9223372036854775808),
+                    ),
+                    items=None,
+                ),
+            ),
+            (
+                _adapter_config(
+                    name="param", required=True, value_type="str", format=FormatAdapter(enum=["TYPE_1", "TYPE_2"])
+                ),
+                ExpectModelProps(
+                    name="param",
+                    required=True,
+                    value_type="str",
+                    format=Format(strategy=FormatStrategy.FROM_ENUMS, enums=["TYPE_1", "TYPE_2"]),
+                    items=None,
+                ),
+            ),
+            (
+                _adapter_config(
+                    name="param",
+                    required=True,
+                    value_type="str",
+                    format=FormatAdapter(formatter=ApiDocValueFormat.DateTime),
+                ),
+                ExpectModelProps(
+                    name="param",
+                    required=True,
+                    value_type="str",
+                    format=Format(
+                        strategy=FormatStrategy.CUSTOMIZE,
+                        customize="datetime_value",
+                        variables=[Variable(name="datetime_value", value_format=ValueFormat.DateTime)],
+                    ),
+                    items=None,
+                ),
+            ),
+            (
+                _adapter_config(
+                    name="array_param",
+                    required=True,
+                    value_type="list",
+                    format=None,
+                    items=[
+                        _adapter_config(
+                            required=True,
+                            value_type="str",
+                            format=FormatAdapter(formatter=ApiDocValueFormat.DateTime),
+                        ),
+                    ],
+                ),
+                ExpectModelProps(
+                    name="array_param",
+                    required=True,
+                    value_type="list",
+                    format=None,
+                    items=[
+                        ExpectModelProps(
+                            name="",
+                            required=True,
+                            value_type="str",
+                            format=Format(
+                                strategy=FormatStrategy.CUSTOMIZE,
+                                customize="datetime_value",
+                                variables=[Variable(name="datetime_value", value_format=ValueFormat.DateTime)],
+                            ),
+                            items=None,
+                        ),
+                    ],
+                ),
+            ),
+            (
+                _adapter_config(
+                    name="array_param",
+                    required=True,
+                    value_type="list",
+                    format=None,
+                    items=[
+                        _adapter_config(
+                            name="id",
+                            required=True,
+                            value_type="str",
+                            format=FormatAdapter(formatter=ApiDocValueFormat.Int64),
+                        ),
+                        _adapter_config(
+                            name="date",
+                            required=True,
+                            value_type="str",
+                            format=FormatAdapter(formatter=ApiDocValueFormat.DateTime),
+                        ),
+                    ],
+                ),
+                ExpectModelProps(
+                    name="array_param",
+                    required=True,
+                    value_type="list",
+                    format=None,
+                    items=[
+                        ExpectModelProps(
+                            name="id",
+                            required=True,
+                            value_type="str",
+                            format=Format(
+                                strategy=FormatStrategy.BY_DATA_TYPE,
+                                size=Size(
+                                    max_value=9223372036854775807,
+                                    min_value=-9223372036854775808,
+                                ),
+                            ),
+                            items=None,
+                        ),
+                        ExpectModelProps(
+                            name="date",
+                            required=True,
+                            value_type="str",
+                            format=Format(
+                                strategy=FormatStrategy.CUSTOMIZE,
+                                customize="datetime_value",
+                                variables=[Variable(name="datetime_value", value_format=ValueFormat.DateTime)],
+                            ),
+                            items=None,
+                        ),
+                    ],
+                ),
+            ),
+        ],
+    )
+    def test_convert_flow(
+        self, adapter_model_obj, pymock_model: BaseProperty, under_test_data: dict, expect: ExpectModelProps
+    ):
+        adapter_model = adapter_model_obj(**under_test_data)
+
+        serialized_data = adapter_model.serialize()
+        pymock_model_has_data = pymock_model.deserialize(serialized_data)
+
+        self._verify_data_model_props(pymock_model=pymock_model_has_data, expect=expect)
+
+    def _verify_data_model_props(self, pymock_model: BaseProperty, expect: ExpectModelProps) -> None:
+        assert pymock_model is not None
+        if expect.name:
+            assert pymock_model.name == expect.name
+        assert pymock_model.required == expect.required
+        assert pymock_model.value_type == expect.value_type
+        assert pymock_model.value_format == expect.format
+        if expect.items:
+            assert pymock_model.items
+            assert len(pymock_model.items) == len(expect.items)
+            # check details of items
+            for pm, e in zip(pymock_model.items, expect.items):
+                self._verify_data_model_props(pymock_model=pm, expect=e)
